# Contabilidad/Contabilidad/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contabilidad",
    "version": "v1"
  },
  "paths": {
...
    "/api/accountingEntry": {
      "get": {
        "tags": [
          "AccountingEntry"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AccountingEntry"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AccountingEntry"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccountingEntry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/accountingEntry/{id}": {
      "get": {
        "tags": [
          "AccountingEntry"
        ],
        "operationId": "GetAccountingEntry",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AccountingEntry"
            }
          }
        }
      },
      "put": {
        "tags": [
          "AccountingEntry"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccountingEntry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "AccountingEntry"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AuxiliarSystem": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "state": {
          "type": "boolean"
        },
        "accountingEntry": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountingEntry"
          }
        }
      }
    },
    "AccountingEntry": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "idAuxiliarSystem": {
          "format": "int32",
          "type": "integer"
        },
        "auxiliarSystem": {
          "$ref": "#/definitions/AuxiliarSystem"
        },
        "account": {
          "type": "string"
        },
        "movementType": {
          "type": "string"
        },
        "entryDate": {
          "format": "date-time",
          "type": "string"
        },
        "seatAmount": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "type": "boolean"
        }
      }
    },
...
  }
}